Add test for checking sequences length
